+++ src/pages/sitemap.xml.ts
@@ -6,11 +6,7 @@ import { GetServerSideProps } from "next";
 // Alternative to host, host is more dynamic and adaptable than hardcoding the url
 // const BASE_URL = "https://cancermodels.org";
 
-function generateSiteMap(
-	modelProviderMixes: { data_source: string; external_model_id: string }[],
-	providers: string[],
-	host: string
-): string {
+function generateSiteMap(providers: string[], host: string): string {
 	return `<?xml version="1.0" encoding="UTF-8"?>
     <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
       ${routes
@@ -51,14 +47,6 @@ function generateSiteMap(
             </url>`
 				)
 				.join("")}
-      ${modelProviderMixes
-				.map(
-					(modelproviderMix) =>
-						`<url>
-              <loc>${host}/data/models/${modelproviderMix.data_source}/${modelproviderMix.external_model_id}</loc>
-            </url>`
-				)
-				.join("")}
     </urlset>
  `;
 }
@@ -68,24 +56,14 @@ function SiteMap() {
 }
 
 export const getServerSideProps: GetServerSideProps = async ({ res, req }) => {
-	// We make API calls to gather the URLs for our site
 	const providersDirectory = path.join(
 		process.cwd(),
 		"/public/static/providers"
 	);
 	const providers = await fs.readdir(providersDirectory);
 
-	const modelProviderMixesRequest = await fetch(
-		`${process.env.NEXT_PUBLIC_API_URL}/search_index?select=external_model_id,data_source`
-	);
-	const modelProviderMixes = await modelProviderMixesRequest.json();
-
-	// We generate the XML sitemap with the models data
-	const sitemap = generateSiteMap(
-		modelProviderMixes,
-		providers,
-		req.headers.host as string
-	);
+	// We generate the XML sitemap with the fetched data
+	const sitemap = generateSiteMap(providers, req.headers.host as string);
 
 	res.setHeader("Content-Type", "text/xml");
 	// we send the XML to the browser
